+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New Site Health "status" sections about OPcache and object cache. 
 - New Site Health "status" section about i18n extension for non `en_US` sites.
 - New Site Health "info" sections about OPcache and object cache. 
-- New Site Health "info" section about the plugin itself. 
+- New Site Health "info" section about the plugin itself.
+- New Site Health "info" section about loggers settings. 
 ### Changed
 - Upgraded Monolog library from version 2.0.1 to version 2.0.2.
 - The SQL listener now generates multiple critical errors if there's more than one SQL error during page rendering.